feat(form): use custom ui:padding prop instead of custom template for schema ui settings
