+++ src/forms/Form.less
@@ -162,20 +162,13 @@
     }
   }
 
-  .propKeyCollapse
-    .ant-collapse-content-box
-    fieldset
-    .ant-col
-    .ant-col:last-of-type
-    .ant-form-item:last-of-type {
-    margin-bottom: 0;
-  }
-
-  .propKeyCollapse
-    .ant-collapse-content-box
-    > .form-group:last-of-type
-    > .ant-form-item:last-of-type {
-    margin-bottom: 0;
+  .propKeyCollapse .ant-collapse-content-box {
+    fieldset .ant-col .ant-col:last-of-type .ant-form-item:last-of-type {
+      margin-bottom: 0;
+    }
+    > .form-group:last-of-type > .ant-form-item:last-of-type {
+      margin-bottom: 0;
+    }
   }
 }
