--- src/forms/templates/UiOptionsObjectFieldTemplate.jsx
@@ -1,7 +1,0 @@
-import ObjectFieldTemplate from "./ObjectFieldTemplate";
-
-const UiOptionsObjectFieldTemplate = (props) => {
-  return <ObjectFieldTemplate {...props} isUiOptions />;
-};
-
-export default UiOptionsObjectFieldTemplate;
